requirements.txt
langchain 
langchain-community 
langchain-openai 
langchain-google-genai 
google-generativeai 
chromadb 
pypdf 
streamlit 
python-dotenv 
tiktoken 
langchain-text-splitters 
